requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
httpx==0.25.2
openai==1.3.7
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6